# Conflicts:
#	app/src/main/java/com/verygoodsecurity/demoapp/MainActivity.kt
#	vgscollect/src/version.gradle

+++ vgscollect/src/main/java/com/verygoodsecurity/vgscollect/view/internal/CVCInputField.kt
@@ -46,7 +46,7 @@ internal class CVCInputField(context: Context): BaseInputField(context) {
     }
 
     override fun dispatchDependencySetting(dependency: Dependency) {
-        val filterLength = InputFilter.LengthFilter(dependency.value)
+        val filterLength = InputFilter.LengthFilter(dependency.value as Int)
         filters = arrayOf(CVCValidateFilter(), filterLength)
         text = text
     }
